# [WIP] Further improve performance of Memory<T>.Span property getter