--- a/src/System.Private.CoreLib/shared/System/Buffers/MemoryManager.cs
+++ b/src/System.Private.CoreLib/shared/System/Buffers/MemoryManager.cs
@@ -3,6 +3,7 @@
 // See the LICENSE file in the project root for more information.
 
 using System.Runtime.CompilerServices;
+using System.Runtime.InteropServices;
 
 namespace System.Buffers
 {
@@ -11,6 +12,19 @@ namespace System.Buffers
     /// </summary>
     public abstract class MemoryManager<T> : IMemoryOwner<T>, IPinnable
     {
+#pragma warning disable CS0414
+        // It's imperative that this be the first field in the MemoryManager<T> instance
+        // and that it have a negative value. Because it's the first field in the object,
+        // it occupies the same position as String.Length and SzArray.Length; but because
+        // this value is negative and those values are not, we can easily determine whether
+        // any given Memory<T> is backed by a variable-length object (string, array) or
+        // a fixed-length object (MemoryManager).
+        private readonly int _dummy = -1;
+#if BIT64
+        private readonly int _dummyPadding = 0;
+#endif
+#pragma warning restore CS0414
+
         /// <summary>
         /// Returns a <see cref="System.Memory{T}"/>.
         /// </summary>
@@ -20,6 +34,24 @@ namespace System.Buffers
         /// Returns a span wrapping the underlying memory.
         /// </summary>
         public abstract Span<T> GetSpan();
+
+        /// <summary>
+        /// A special helper method which serves as a stub for the call to GetSpan(). It's
+        /// non-virtual and forbids inlining to keep the call site as compact as possible.
+        /// Additionally, since the span is deconstructed rather than returned by value,
+        /// it means the caller only has to allocate space for a single Int32 on the stack.
+        /// </summary>
+        [MethodImpl(MethodImplOptions.NoInlining)]
+        internal ByReference<T> GetSpanAndDeconstruct(out int spanLength)
+        {
+            // Returns ByReference<T> because the C# compiler sees the 'out' parameter and
+            // reasons that returning 'ref T' might lead to a reference escaping a local
+            // scope. The ByReference<T> return type prevents this false positive error.
+
+            Span<T> span = GetSpan();
+            spanLength = span.Length;
+            return new ByReference<T>(ref MemoryMarshal.GetReference(span));
+        }
 
         /// <summary>
         /// Returns a handle to the memory that has been pinned and hence its address can be taken.