Working through further experiments with minimizing the overhead of the `{ReadOnly}Memory<T>.Span` property getter. Trying to minimize the number of comparisons performed, minimize the stack space required, and utilize register allocations more effectively.

Results from an x64 run are below. I didn't perform an x86 run yet.

|                       Method |     Toolchain |       Mean |     Error |    StdDev | Ratio | RatioSD |
|----------------------------- |-------------- |-----------:|----------:|----------:|------:|--------:|
|  GetSpan_MemOfBytesFromArray |  3.0-preview3 | 1,955.4 us | 26.541 us | 23.528 us |  1.00 |    0.00 |
|  GetSpan_MemOfBytesFromArray | exp. 27423-58 | 1,708.6 us | 12.064 us | 10.074 us |  0.87 |    0.01 |
|                              |               |            |           |           |       |         |
| GetSpan_MemOfBytesFromMemMgr |  3.0-preview3 | 3,132.1 us | 38.245 us | 31.937 us |  1.00 |    0.00 |
| GetSpan_MemOfBytesFromMemMgr | exp. 27423-58 | 4,304.9 us | 50.077 us | 46.842 us |  1.38 |    0.02 |
|                              |               |            |           |           |       |         |
|      GetSpan_MemOfBytesEmpty |  3.0-preview3 | 1,128.9 us |  7.760 us |  6.879 us |  1.00 |    0.00 |
|      GetSpan_MemOfBytesEmpty | exp. 27423-58 |   860.6 us |  5.582 us |  5.222 us |  0.76 |    0.01 |
|                              |               |            |           |           |       |         |
|  GetSpan_MemOfCharsFromArray |  3.0-preview3 | 2,364.8 us | 36.351 us | 32.225 us |  1.00 |    0.00 |
|  GetSpan_MemOfCharsFromArray | exp. 27423-58 | 1,788.2 us | 32.345 us | 27.010 us |  0.76 |    0.02 |
|                              |               |            |           |           |       |         |
| GetSpan_MemOfCharsFromString |  3.0-preview3 | 1,791.3 us | 24.862 us | 23.256 us |  1.00 |    0.00 |
| GetSpan_MemOfCharsFromString | exp. 27423-58 | 1,642.9 us | 12.501 us | 11.694 us |  0.92 |    0.01 |
|                              |               |            |           |           |       |         |
| GetSpan_MemOfCharsFromMemMgr |  3.0-preview3 | 3,152.8 us | 23.027 us | 21.540 us |  1.00 |    0.00 |
| GetSpan_MemOfCharsFromMemMgr | exp. 27423-58 | 4,573.2 us | 64.132 us | 59.989 us |  1.45 |    0.02 |
|                              |               |            |           |           |       |         |
|      GetSpan_MemOfCharsEmpty |  3.0-preview3 | 1,427.8 us | 11.145 us |  8.701 us |  1.00 |    0.00 |
|      GetSpan_MemOfCharsEmpty | exp. 27423-58 |   923.9 us | 18.260 us | 16.187 us |  0.65 |    0.01 |

Codegen for the x64 implementation of the `ReadOnlyMemory<char>.Span` property getter is also around 20% smaller: 118 bytes before, 95 bytes after.

The regression in the _from `MemoryManager<T>`_ runs isn't surprising due to the extra call now required here. This was a deliberate trade-off: if `MemoryManager<T>` usage is expected to be rare, deprioritize it and optimize the codegen for the more common cases. Since we already made improvements in this code path between 2.1 and 3.0 (see https://github.com/dotnet/coreclr/pull/20386), we're still within around 5% of the 2.1 performance numbers.

Marked WIP because it's not fully implemented (haven't touched `Memory<T>.Span` yet) or tested.