@stephentoub 2019-02-25
> if MemoryManager<T> usage is expected to be rare

I'm not convinced it is. This is the only way today to use a pointer/length with Memory, and we're already seeing developers relying on that.

It of course still might be worth a trade-off here, but we should still aim not to regress it.