@GrabYourPitchforks 2019-03-25
@jkotas @stephentoub Is it worthwhile for me to experiment with a change where we can make all three scenarios faster, but at the cost of increasing the size of `MemoryManager<T>` by `sizeof(IntPtr)`? Or is increasing the size of that type so unpalatable as to make this a non-starter?