@jkotas 2019-03-25
I do not think that this optimization is worth looking into.

I think there needs to be a ballance between performance and understandability/maintanability of the code. The fake length field does not strike the right ballance for me.