@GrabYourPitchforks 2019-03-27
Thanks both for the feedback! I'm going to go ahead and close this since the Length trick really was the primary mechanism to increase perf, and if that doesn't sit well with folks then this probably isn't worth pursuing further.